Added instrumentArray to App component as well as cleaned up some code and added notes to some of the functionality
