+++ src/pages/Arts/Arts.js
@@ -2,5 +2,5 @@ import React from "react";
 import { ButtonList } from "components";
 
 export default function Math(props) {
-  return <ButtonList list={props.numbers} category="math" />;
+  return <ButtonList list={props.instruments} category="arts" />;
 }
