+++ src/pages/Science/Science.js
@@ -2,5 +2,5 @@ import React from "react";
 import { ButtonList } from 'components';
 
 export default function Science(props) {
-  return <ButtonList list={props.alphabet} category="science" />;
+  return <ButtonList list={props.animals} category="science" />;
 }
